Basic whitelist with add feature.
The basic structure for the whitelist is now working, but only with the add feature, and a few UI bugs. I tested with developer.mozilla.org (https://developer.mozilla.org/samples/xultu/examples/aligner.xul), and the test appeared to work, although the scripts on the XUL page failed to load for some reason.