Fixed domain string and UI bugs.
The domain was a little odd to manage because internally only the host name is stored, while the input requires the full domain, with the protocol. I just added a function that adds the protocol when needed, so now in theory the user only needs to handle hosts.

Also fixed the bug that caused the window to lose the OK button when the first host was added.

Finally, the window is now focused when attempting to open it if it's already open.