Added support for local files (file://).
Support for local files was surprisingly complicated. There are some oddities that happen right after changing the DB that I couldn't overcome. In the end it was just easier to ask the user to restart right away.

Also fixed the log level. It's now down to Warn.

Upped version number to 0.6. Not sure if I'll release this version or add more stuff before. Depends on whether 0.5 is approved on AMO soon.