First working installer generator.
This code generates a Remote XUL installer that is restarless. On installation, it configures the remote XUL whitelist and then removes itself. It's easy to modify from the config.mk file.
I haven't tested this heavily, it's just the basis for the code that will be included in the add-on.