Bug fixes and polish on generator feature.
Changed the UI in the generator to display the default texts (hardcoded) and provide more room to edit the longer messages.
Fixed a bug in the installer generation that prevented anything to be edited individually, except for the title.
Added trim calls to the installer inputs to prevent whitespace issues.