Fixed local file migration code for manager and import.
This now works for the main manager window and the import/export feature. In the end it was best not to have a separate preference to handle the migration. The add-on will have to run a query at startup for some time, but the whole code can be removed later on.

Also fixed the Makefile to adjust for packed add-ons (whoops).